+++ backend/driver.py
@@ -66,8 +66,8 @@ def _generate_launcher(constants, signature, kernel_name):
 #include <assert.h>
 #include <stdbool.h>
 #include <Python.h>
-#include "CRunnerUtils.h"
-#include "CRunnerUtils.cpp"
+#include "ExecutionEngine/CRunnerUtils.h"
+#include "ExecutionEngine/CRunnerUtils.cpp"
 
 extern "C" {{
   // Pointer type (=Memref) becomes int64_t + MemRef struct
@@ -197,7 +197,7 @@ def compile_module(launcher_src, kernel_placeholder_name):
         scheme = 'posix_prefix'
     py_include_dir = sysconfig.get_paths(scheme=scheme)["include"]
     cpu_backend_path = Path(__file__).resolve().parent
-    execution_engine_include_dir = os.path.join(cpu_backend_path, "ExecutionEngine")
+    include_dir = os.path.join(cpu_backend_path, "include")
 
     def launch(
         gridX, gridY, gridZ, num_warps, num_ctas, clusterDim0, clusterDim1, clusterDim2,
@@ -222,7 +222,7 @@ def compile_module(launcher_src, kernel_placeholder_name):
               Path(asm_src_path).write_bytes(asm_src)
               Path(launcher_src_path).write_text(src)
               # Compile it together.
-              subprocess.check_call(["g++", launcher_src_path, asm_src_path, f"-I{py_include_dir}", f"-I{execution_engine_include_dir}", "-shared", "-fPIC", "-o", so_path])
+              subprocess.check_call(["g++", launcher_src_path, asm_src_path, f"-I{py_include_dir}", f"-I{include_dir}", "-shared", "-fPIC", "-o", so_path])
 
               with open(so_path, "rb") as f:
                 cache_path = cache.put(f.read(), filename, binary=True)
@@ -267,7 +267,7 @@ class CPUUtils(object):
     # the triton kernels.
     # Since we recompile the kernel every time (see compile_module above),
     # and the metadata generated by these functions aren't applicable to the cpu
-    # backend, just define expose the same functions with dummy implementation.
+    # backend, just define the same functions with dummy implementation.
     @staticmethod
     def get_device_properties(device):
         return {
