--- python/ExecutionEngine/version.txt
@@ -1,1 +1,0 @@
-https://github.com/llvm/llvm-project/commit/3be3883e6d67bf908fd12b51219075293ebb3dff
